Fixing warnings Re #966 and #3653

+++ Code/Mantid/Framework/API/src/CompositeFunctionMW.cpp
@@ -63,7 +63,7 @@ std::string CompositeFunctionMW::asString()const
     }
   }
   std::string ties;
-  for(int i=0;i<nParams();i++)
+  for(size_t i=0;i<nParams();i++)
   {
     const ParameterTie* tie = getTie(i);
     if (tie)
